refactor: Update routing and guard for select-patient page
Update the routing configuration in the app-routing.module.ts file to use the VerifyRegistrationComponent as the component for the 'verify-registration' path. Additionally, update the canActivate guard to use the patientSelectedGuard instead of the registrationSelectedGuard.

This change is necessary to ensure that a patient is selected before accessing the 'verify-registration' page.